Major refactoring
Move encoding and language options to the main .tex file where they are
commonly placed. Separate the preamble directives inherent to the
bachelor work into a separate file from the preamble. Now `bachelor.tex`
is a drop-in file for specifying the design, and `preamble.tex`
accompanies `example.tex` in being an exemplary set of common, not
necessarily mandatory, tricks.